modules/tprimat/amd-requirements.txt
torch>=2.0.0
lightning>=2.0.0

matplotlib>=3.5.0
seaborn>=0.12.0
numpy>=1.21.0
pyyaml>=6.0

datasets>=2.10.0
huggingface_hub>=0.13.0
transformers>=4.30.0

tensorboard>=2.10.0
